requirements.txt
<<<<<<< HEAD
alembic==1.14.1
amqp==5.3.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
astroid==3.3.8
billiard==4.2.1
black==25.1.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
cssselect2==0.7.0
dill==0.3.9
Flask==3.1.0
Flask-DebugToolbar==0.16.0
Flask-DotEnv==0.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.8
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.2
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pylint==3.3.4
pyparsing==3.2.1
pypdf==5.2.0
pyphen==0.17.2
pytest==8.3.4
pytest-flask==1.3.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
reportlab==4.3.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.37
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==64.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
=======
authlib
requests
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
>>>>>>> 90f5ddd1 (Initial commit: Integrated Google Classroom API and report generation)
